+++ includes/settings/mybooking-plugin-settings.php
@@ -417,6 +417,13 @@ EOF;
 		                     'mybooking-plugin-configuration',
 		                     'mybooking_plugin_settings_css');
 
+		  add_settings_field('mybooking_plugin_settings_components_js_bs_modal_backdrop_compatibility',
+		                     'Boostrap Modal backdrop',
+		                     array($this, 'field_mybooking_plugin_settings_components_js_bs_modal_backdrop_compatibility_callback'),
+		                     'mybooking-plugin-configuration',
+		                     'mybooking_plugin_settings_css');
+
+
 		}
 
     // == Connection
@@ -910,6 +917,30 @@ EOF;
 		  echo "<p class=\"description\">Use <b>Bootstrap modal no conflict</b> if there is a conflict with Bootstrap modal."; 
 		  echo "<p class=\"description\">It will use $('#my_modal').bootstrapModal('show') instead of $('#my_modal').modal('show').</p>";
 		  echo "<p class=\"description\">It only applies if you are using the plugin custom CSS/JS Framework.</p>";
+
+		}
+
+		/**
+		 * Render Mybooking Bootstrap Modal Backdrop
+		 */
+		public function field_mybooking_plugin_settings_components_js_bs_modal_backdrop_compatibility_callback() {
+		  
+		  $settings = (array) get_option("mybooking_plugin_settings_css");
+		  $field = "mybooking_plugin_settings_components_js_bs_modal_backdrop_compatibility";
+		  if (array_key_exists($field, $settings)) {
+		    $value = esc_attr( $settings[$field] );
+		  }
+		  else {
+        $value = '';
+		  } 
+		  
+		  $checked = ($value == '1') ? 'checked' : '';
+      echo "<input type='hidden' name='mybooking_plugin_settings_css[$field]' value=''/>";
+		  echo "<input type='checkbox' name='mybooking_plugin_settings_css[$field]' value='1' $checked class='regular-text' />";
+
+		  echo "<p class=\"description\">Setup <b>Bootstrap modal backdrop compatibility</b> when showing a modal."; 
+		  echo "<p class=\"description\">It helps to improve plugin compatibility with some themes when the modal is not shown on top.</p>";
+		  echo "<p class=\"description\">Use only if reservation process modal is show behind the backdrop and components are not clickable.</p>";
 
 		}